add auto-complete for command line arguments

+++ internal/commands/macbaremetal/security_group_rule.go
@@ -56,14 +56,23 @@ func (s *securityGroupRuleListCommand) Run(cmd *cobra.Command, args []string) er
 	return commands.PrintStdout(items)
 }
 
+func (s *securityGroupRuleListCommand) CompleteArg(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	if len(args) == 0 {
+		return completeSecurityGroup(cmd.Context(), toComplete)
+	}
+
+	return nil, cobra.ShellCompDirectiveNoFileComp
+}
+
 func (s *securityGroupRuleListCommand) Build() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "list SECURITY-GROUP",
-		Aliases: []string{"show", "ls", "get"},
-		Short:   "List security group rules",
-		Long:    "Lists all mac bare metal security group rules.",
-		Args:    cobra.ExactArgs(1),
-		RunE:    s.Run,
+		Use:               "list SECURITY-GROUP",
+		Aliases:           []string{"show", "ls", "get"},
+		Short:             "List security group rules",
+		Long:              "Lists all mac bare metal security group rules.",
+		Args:              cobra.ExactArgs(1),
+		ValidArgsFunction: s.CompleteArg,
+		RunE:              s.Run,
 	}
 
 	cmd.Flags().StringVar(&s.filter, "filter", "", "custom term to filter the results")
@@ -112,13 +121,20 @@ func (s *securityGroupRuleCreateCommand) Run(cmd *cobra.Command, args []string) 
 	return commands.PrintStdout(item)
 }
 
+func (s *securityGroupRuleCreateCommand) CompleteArg(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	if len(args) == 0 {
+		return completeSecurityGroup(cmd.Context(), toComplete)
+	}
+
+	return nil, cobra.ShellCompDirectiveNoFileComp
+}
+
 func (s *securityGroupRuleCreateCommand) Build() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "create SECURITY-GROUP",
 		Aliases: []string{"add", "new"},
 		Short:   "Create new security group",
 		Long:    "Creates a new mac bare metal security group.",
-		Args:    cobra.ExactArgs(1),
 		Example: commands.FormatExamples(fmt.Sprintf(`
 			# Create rule to allow tcp traffic on port 80 (HTTP) from any source IP
 			%[1]s mac-bare-metal security-group rule create default --direction ingress --protocol tcp --from-port 80 --to-port 80
@@ -126,7 +142,9 @@ func (s *securityGroupRuleCreateCommand) Build() *cobra.Command {
 			# Create rule to allow tcp traffic on port 22 (SSH) only from subnet 1.1.1.0/24
 			%[1]s mac-bare-metal security-group rule create default --direction ingress --protocol tcp --from-port 22 --to-port 22 --ip-range 1.1.1.0/24
 		`, commands.Name)),
-		RunE: s.Run,
+		Args:              cobra.ExactArgs(1),
+		ValidArgsFunction: s.CompleteArg,
+		RunE:              s.Run,
 	}
 
 	cmd.Flags().StringVar(&s.direction, "direction", "", "direction of the rule")
@@ -194,18 +212,35 @@ func (s *securityGroupRuleUpdateCommand) Run(cmd *cobra.Command, args []string) 
 	return commands.PrintStdout(item)
 }
 
+func (s *securityGroupRuleUpdateCommand) CompleteArg(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	if len(args) == 0 {
+		return completeSecurityGroup(cmd.Context(), toComplete)
+	}
+
+	if len(args) == 1 {
+		securityGroup, err := findSecurityGroup(cmd.Context(), args[0])
+		if err != nil {
+			return nil, cobra.ShellCompDirectiveError
+		}
+
+		return completeSecurityGroupRule(cmd.Context(), securityGroup, toComplete)
+	}
+
+	return nil, cobra.ShellCompDirectiveNoFileComp
+}
+
 func (s *securityGroupRuleUpdateCommand) Build() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "update SECURITY-GROUP RULE",
 		Short: "Update security group rule",
 		Long:  "Updates a mac bare metal security group rule.",
-
 		Example: commands.FormatExamples(fmt.Sprintf(`
 			# Update SSH rule to allow tcp traffic from broader subnet 1.1.0.0/16
 			%[1]s mac-bare-metal security-group rule update default 1234 --direction ingress --protocol tcp --from-port 22 --to-port 22 --ip-range 1.1.0.0/16
 		`, commands.Name)), // TODO
-		Args: cobra.ExactArgs(2),
-		RunE: s.Run,
+		Args:              cobra.ExactArgs(2),
+		ValidArgsFunction: s.CompleteArg,
+		RunE:              s.Run,
 	}
 
 	cmd.Flags().StringVar(&s.direction, "direction", "", "direction of the rule")
@@ -244,7 +279,10 @@ func (s *securityGroupRuleDeleteCommand) Run(cmd *cobra.Command, args []string) 
 		return fmt.Errorf("find security group rule: %w", err)
 	}
 
-	// TODO ask for confirmation
+	if !s.force && !commands.ConfirmDeletion("security group", securityGroup) {
+		commands.Stderr.Println("aborted.")
+		return nil
+	}
 
 	err = service.Delete(cmd.Context(), rule.ID)
 	if err != nil {
@@ -254,29 +292,49 @@ func (s *securityGroupRuleDeleteCommand) Run(cmd *cobra.Command, args []string) 
 	return nil
 }
 
+func (s *securityGroupRuleDeleteCommand) CompleteArg(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	if len(args) == 0 {
+		return completeSecurityGroup(cmd.Context(), toComplete)
+	}
+
+	if len(args) == 1 {
+		securityGroup, err := findSecurityGroup(cmd.Context(), args[0])
+		if err != nil {
+			return nil, cobra.ShellCompDirectiveError
+		}
+
+		return completeSecurityGroupRule(cmd.Context(), securityGroup, toComplete)
+	}
+
+	return nil, cobra.ShellCompDirectiveNoFileComp
+}
+
 func (s *securityGroupRuleDeleteCommand) Build() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "delete SECURITY-GROUP RULE",
-		Aliases: []string{"del", "remove", "rm"},
-		Short:   "Delete security group rule",
-		Long:    "Deletes a mac bare metal security group rule.",
-		Args:    cobra.ExactArgs(2),
-		RunE:    s.Run,
+		Use:               "delete SECURITY-GROUP RULE",
+		Aliases:           []string{"del", "remove", "rm"},
+		Short:             "Delete security group rule",
+		Long:              "Deletes a mac bare metal security group rule.",
+		Args:              cobra.ExactArgs(2),
+		ValidArgsFunction: s.CompleteArg,
+		RunE:              s.Run,
 	}
 
 	return cmd
 }
 
-func findSecurityGroup(ctx context.Context, term string) (macbaremetal.SecurityGroup, error) {
-	securityGroups, err := macbaremetal.NewSecurityGroupService(commands.Config.Client).List(ctx)
+func completeSecurityGroupRule(ctx context.Context, securityGroup macbaremetal.SecurityGroup, term string) ([]string, cobra.ShellCompDirective) {
+	rules, err := macbaremetal.NewSecurityGroupRuleService(commands.Config.Client, securityGroup.ID).List(ctx)
 	if err != nil {
-		return macbaremetal.SecurityGroup{}, fmt.Errorf("fetch security groups: %w", err)
+		return nil, cobra.ShellCompDirectiveError
 	}
 
-	securityGroup, err := filter.FindOne(securityGroups, term)
-	if err != nil {
-		return macbaremetal.SecurityGroup{}, fmt.Errorf("find security group: %w", err)
+	filtered := filter.Find(rules, term)
+
+	names := make([]string, len(filtered))
+	for i, rule := range filtered {
+		names[i] = fmt.Sprint(rule.ID)
 	}
 
-	return securityGroup, nil
+	return names, cobra.ShellCompDirectiveNoFileComp
 }
